requirements.txt
architect-functions
boto3
Jinja2
weasyprint
requests
python-json-logger
pyjwt